ai-service/requirements.txt
flask
flask-cors
flask-limiter
groq
chromadb
sentence-transformers
python-dotenv
cachetools